print fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned long long int  n1, n2, next;
+	double  n1, n2, next;
 	int i;
 
 	n1 = 1;
@@ -15,10 +15,10 @@ int main(void)
 	next = n1 + n2;
 	i = 3;
 
-	printf("%llu, %llu ", n1, n2);
+	printf("%.0f, %.0f ", n1, n2);
 	while (i < 98)
 	{
-		printf("%llu, ", next);
+		printf("%.0f, ", next);
 		n1 = n2;
 		n2 = next;
 		next = n1 + n2;
